requirements.txt
pillow
numpy
torch
torchvision
opencv-contrib-python